django_project_491/requirements.txt
Django==5.1.2
python-dotenv==1.0.1
requests==2.32.3
SPARQLWrapper==2.0.0
django-cors-headers==4.0.0
gunicorn==22.0.0
pymysql==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.17
pillow==11.0.0